+++ CHANGELOG.md
@@ -1,4 +1,15 @@
 
+TODO
+- allow filters ? (prevent null or duplicated data)
+- better subjects
+- more operators
+- more complex case
+
+(next)
+------
+- null & undefined now allowed as static values
+- improved log traces
+
 v0.0.4
 ------
 - updated API to allow more custom build of the generator by passing dependencies as param